# openapi.committed.json
{
  "components": {
    "schemas": {
...
      "CreateReportRequest": {
        "properties": {
          "emotional_state": {
            "default": "neutral",
            "title": "Emotional State",
            "type": "string"
          },
          "execution_quality": {
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "F"
            ],
            "title": "Execution Quality",
            "type": "string"
          },
          "followed_plan": {
            "title": "Followed Plan",
            "type": "boolean"
          },
          "lessons_learned": {
            "default": "",
            "title": "Lessons Learned",
            "type": "string"
          },
          "setup_quality": {
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "F"
            ],
            "title": "Setup Quality",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "setup_quality",
          "execution_quality",
          "followed_plan"
        ],
        "title": "CreateReportRequest",
        "type": "object"
      },
      "CreateTradeRequest": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "action": {
            "title": "Action",
            "type": "string"
          },
          "commission": {
            "default": 0.0,
            "title": "Commission",
            "type": "number"
          },
          "exec_id": {
            "title": "Exec Id",
            "type": "string"
          },
          "instrument": {
            "title": "Instrument",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "realized_pnl": {
            "default": 0.0,
            "title": "Realized Pnl",
            "type": "number"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "exec_id",
          "time",
          "instrument",
          "action",
          "quantity",
          "price",
          "account_id"
        ],
        "title": "CreateTradeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateReportRequest": {
        "properties": {
          "emotional_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotional State"
          },
          "execution_quality": {
            "anyOf": [
              {
                "enum": [
                  "A",
                  "B",
                  "C",
                  "D",
                  "F"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Quality"
          },
          "followed_plan": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followed Plan"
          },
          "lessons_learned": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lessons Learned"
          },
          "setup_quality": {
            "anyOf": [
              {
                "enum": [
                  "A",
                  "B",
                  "C",
                  "D",
                  "F"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Quality"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "UpdateReportRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Trading journal REST API",
    "title": "Zorivest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trades": {
      "get": {
        "description": "List trades with optional account filter, search, and sort.",
        "operationId": "list_trades_api_v1_trades_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-time",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Trades",
        "tags": [
          "trades"
        ]
      },
      "post": {
        "description": "Create a new trade.",
        "operationId": "create_trade_api_v1_trades_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Trade",
        "tags": [
          "trades"
        ]
      }
    },
...
    "/api/v1/trades/{exec_id}/report": {
      "delete": {
        "description": "Delete the report for a trade.",
        "operationId": "delete_report_api_v1_trades__exec_id__report_delete",
        "parameters": [
          {
            "in": "path",
            "name": "exec_id",
            "required": true,
            "schema": {
              "title": "Exec Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Report",
        "tags": [
          "reports"
        ]
      },
      "get": {
        "description": "Get the report for a trade.",
        "operationId": "get_report_api_v1_trades__exec_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "exec_id",
            "required": true,
            "schema": {
              "title": "Exec Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "reports"
        ]
      },
      "post": {
        "description": "Create a post-trade report.",
        "operationId": "create_report_api_v1_trades__exec_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "exec_id",
            "required": true,
            "schema": {
              "title": "Exec Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report",
        "tags": [
          "reports"
        ]
      },
      "put": {
        "description": "Update the report for a trade.",
        "operationId": "update_report_api_v1_trades__exec_id__report_put",
        "parameters": [
          {
            "in": "path",
            "name": "exec_id",
            "required": true,
            "schema": {
              "title": "Exec Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Report",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
...
}